[core] Merge SpriteStore and SpriteAtlas
