+++ src/mbgl/style/observer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <mbgl/text/glyph_store_observer.hpp>
-#include <mbgl/sprite/sprite_store_observer.hpp>
+#include <mbgl/sprite/sprite_atlas_observer.hpp>
 #include <mbgl/style/source_observer.hpp>
 #include <mbgl/map/update.hpp>
 
@@ -9,7 +9,7 @@ namespace mbgl {
 namespace style {
 
 class Observer : public GlyphStoreObserver,
-                 public SpriteStoreObserver,
+                 public SpriteAtlasObserver,
                  public SourceObserver {
 public:
     virtual void onUpdate(Update) {}
