+++ src/mbgl/style/style.hpp
@@ -6,7 +6,7 @@
 #include <mbgl/style/layer_observer.hpp>
 #include <mbgl/style/update_batch.hpp>
 #include <mbgl/text/glyph_store_observer.hpp>
-#include <mbgl/sprite/sprite_store_observer.hpp>
+#include <mbgl/sprite/sprite_atlas_observer.hpp>
 #include <mbgl/map/mode.hpp>
 #include <mbgl/map/zoom_history.hpp>
 
@@ -26,7 +26,6 @@ namespace mbgl {
 class FileSource;
 class GlyphStore;
 class GlyphAtlas;
-class SpriteStore;
 class SpriteAtlas;
 class LineAtlas;
 class RenderData;
@@ -38,7 +37,7 @@ class UpdateParameters;
 class QueryParameters;
 
 class Style : public GlyphStoreObserver,
-              public SpriteStoreObserver,
+              public SpriteAtlasObserver,
               public SourceObserver,
               public LayerObserver,
               public util::noncopyable {
@@ -106,7 +105,6 @@ public:
     FileSource& fileSource;
     std::unique_ptr<GlyphStore> glyphStore;
     std::unique_ptr<GlyphAtlas> glyphAtlas;
-    std::unique_ptr<SpriteStore> spriteStore;
     std::unique_ptr<SpriteAtlas> spriteAtlas;
     std::unique_ptr<LineAtlas> lineAtlas;
