+++ src/mbgl/tile/geometry_tile.cpp
@@ -27,7 +27,6 @@ GeometryTile::GeometryTile(const OverscaledTileID& id_,
       worker(parameters.workerScheduler,
              ActorRef<GeometryTile>(*this, mailbox),
              id_,
-             *parameters.style.spriteStore,
              *parameters.style.glyphAtlas,
              *parameters.style.glyphStore,
              obsolete,
