+++ src/mbgl/tile/geometry_tile_worker.hpp
@@ -14,7 +14,6 @@ namespace mbgl {
 
 class GeometryTile;
 class GeometryTileData;
-class SpriteStore;
 class GlyphAtlas;
 class GlyphStore;
 class SymbolLayout;
@@ -28,7 +27,6 @@ public:
     GeometryTileWorker(ActorRef<GeometryTileWorker> self,
                        ActorRef<GeometryTile> parent,
                        OverscaledTileID,
-                       SpriteStore&,
                        GlyphAtlas&,
                        GlyphStore&,
                        const std::atomic<bool>&,
@@ -49,7 +47,6 @@ private:
     ActorRef<GeometryTile> parent;
 
     const OverscaledTileID id;
-    SpriteStore& spriteStore;
     GlyphAtlas& glyphAtlas;
     GlyphStore& glyphStore;
     const std::atomic<bool>& obsolete;
